requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
idna==3.4
model-bakery==1.12.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
sqlparse==0.4.3
urllib3==2.0.3
